moves UIToaster to loadControl
